api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
groq>=0.4.0
sqlalchemy>=2.0.0
edge-tts>=6.1.0
huggingface-hub>=0.20.0
matplotlib>=3.8.0
